Move Char from atoms.py to stdlib.py
